LibGPU+LibSoftGPU: Move RasterPosition into its own header in LibGPU
